Add UART interrupt slave com and accelerometer
The UART component of the AVR_Lib was extended with a class for
slave-side interrupt driven UART communication. A send request is
triggered by an external interrupt. In that case, a UARTDataSource is
used to collect data to be transmitted to the master. UARTDataSource can
also be used for other UART communication patterns.

The Accelerometer class was completed. It was extended with an
auto-calibrate function. Also, angle conversion was added.